Fix ADC reading and scaling. - Old ADC scale was just completely wrong lol, it just coincidentally looked right cause I was using high-impedance mode on wavegen - Config+Read ADC channels one at a time. Reading them sequentially in discontinuous mode like I was doing before doesn't work since they don't all get read every iteration of the loop

+++ Core/Inc/sensor.h
@@ -4,13 +4,13 @@
 #include <stdint.h>
 
 /*
- * Convert a raw 16-bit ADC value to millivolts.
+ * Convert a raw ADC value to millivolts.
  */
-uint16_t adc_raw_to_mv(const uint16_t raw_value);
+uint16_t adc_raw_to_mv(const uint32_t raw_value);
 
 /*
- * Convert a raw 16-bit ADC value from a pressure transducer (PT) to PSI.
+ * Convert a raw ADC value from a pressure transducer (PT) to PSI.
  */
-uint16_t pt_adc_raw_to_psi(const uint16_t raw_value);
+uint16_t pt_adc_raw_to_psi(const uint32_t raw_value);
 
 #endif /* INJ_SENSOR_HUB_SENSOR_H */
